ytdl_hook.lua: set last url->playlist data as shared property
Previously clients could not detect easily whether ytdl_hook.lua
replaced a URL with a playlist.

Now, whenever such replacement happens, ytdl_hook.lua sets the shared
script property "ytdl_hook-list" to a JSON string of this object:
 {base: <original-url>, list: <playlist-like-array>, time: <mp-time>}

where the value of time is the mpv internal timestamp in seconds
(in a script - mp.get_time()) when the replacement happened, and list
is an array of objects where each object has a "filename" string field
and possibly also a "title" string field.

Clients can check the value of this shared scripts property to examine
the last replacement (or observe, but beware of dummy callbacks,
possibly even before this property was even set).