lua: support custom init via ~~/init.lua (WIP)
This is similar to JS, where if ~~/init.lua exists, then it's invoked
after the builtin defaults.lua and before the actual script is loaded,
so that it can augment the builtin script-environment initialization
with a global init script and apply to all scripts.

This can be used for instance to modify package.path for all scripts,
or add some lua-version compatibility shims, etc.

The script executes in the global (lua) environment but in a new file
context, such that local variables/functions are hidden from other
files (including from the target script which is loaded next), while
non-local identifiers become global and accessible by other scripts
(same as when a JS module is executed).

The custom-init file is ignored if mpv is invoked with --no-config.

TODO: docs, and decide if we actually want it (probably not).